Mark soft hyphens w/ hyphen package
Puppeteer in headless mode isn't capable of automatically hyphenating.
It's "wontfix" as it is intended, see:
https://bugs.chromium.org/p/chromium/issues/detail?id=1261480#c5

On the other hand, Puppeteer can't print PDFs in headful mode.
In the future, perhaps a more "native" headless mode would be the
solution. Refer to the previous bugthread for more detail.

In the meantime, automatically mark soft hyphens for lang:en-gb
programmatically.
Also, disable hyphenation in <main> content, which is wide enough to be
left untouched.