WIP: Use SVG mask to polish timeline graph
There's a problem, however: when rendered at 100% (without zooming in),
it appears that the resulting PDF is not quite right in Chromium.  It is
rendered well in Firefox.  Zooming in in Chromium does remedy this
issue, which means that the problem really is with Chromium's rendering
rather than with the PDF itself at the time of generation.

I tried a couple of things:

- rendering it with all versions of Puppeteer until a version so ancient
  it wouldn't even render SVGs at all (<19.0.0),
- rendering it using free online HTML-to-PDF converters,
- placing the `mask` in its own `defs` at the top-level of the current
  svg...

To no avail.  I haven't thought of a valid solution yet.  Well, I could
consider creating a different mask for each "node", which may possibly
work... But it's pretty lame, especially considering that the maths is
right and all aligns precisely right in principle and that the rendering
issue only resides at the *consumer level*.

Also make its branches thinner: 4px instead of 5.  Using an even number
has the advantage of also presenting a crisper edge when rendering, it
seems.

Additionally round the stroke-linecap.