Further tweak fonts, explicitly use faux bold
Prepare text for use with the "Cinzel Decorative" font.

I want it slightly bolder, but I prefer the "faux bold" version
generated by the browser when the font-face doesn't actually offer a
weightier version.

The "faux bold" computed for Cinzel Decorative is slightly superior
to that for Cinzel, but the decorative elements of -Decorative (which
appear to only exist on the uppercase glyphs) are somewhat too much in
this context.

Therefore, this function transforms all letters in minuscule, then
makes the capital ones slightly bigger, in some sort of a reverse
"font-variant: small-caps".