Remove experiences' tags
There's less room for them now, and they were somewhat unhelpful, as
well as being rounded, which didn't quite adhere to the geometric theme
I adopted.

To make up for their removal, I slightly bumped up the vertical gaps in
each experience article, as well as the equivalent ones in the
highlights on the second page.