Drop .editorconfig
I love the thing. But, ya know, you shouldn't even use [*], as it'd
apply to every single (non-binary?) file in every single sub-directory,
including your build directory, your .git-dir...

Whereas here I've already set up linting for the content that needs it,
and in the end, this is a solo project built by an obsessive guy with a
reasonable editor: I should be fine.